+++ src/pipelines/manifest.rs
@@ -1,4 +1,4 @@
-//! Icon asset pipeline.
+//! Manifest asset pipeline (https://www.w3.org/TR/appmanifest/)
 
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -11,34 +11,28 @@ use super::ATTR_HREF;
 use super::{AssetFile, HashedFileOutput, LinkAttrs, TrunkLinkPipelineOutput};
 use crate::config::RtcBuild;
 
-/// An Icon asset pipeline.
-pub struct Icon {
+/// A manifest asset pipeline.
+pub struct Manifest {
     /// The ID of this pipeline's source HTML element.
     id: usize,
     /// Runtime build config.
     cfg: Arc<RtcBuild>,
     /// The asset file being processed.
     asset: AssetFile,
-    /// Optional type attribute (like type="image/x-icon")
-    icon_type: Option<String>,
-    /// Optional sizes attribute
-    sizes: Option<String>,
 }
 
-impl Icon {
-    pub const TYPE_ICON: &'static str = "icon";
+impl Manifest {
+    pub const TYPE_MANIFEST: &'static str = "manifest";
 
     pub async fn new(cfg: Arc<RtcBuild>, html_dir: Arc<PathBuf>, attrs: LinkAttrs, id: usize) -> Result<Self> {
         // Build the path to the target asset.
         let href_attr = attrs
             .get(ATTR_HREF)
-            .context(r#"required attr `href` missing for <link data-trunk rel="icon" .../> element"#)?;
+            .context(r#"required attr `href` missing for <link data-trunk rel="manifest" .../> element"#)?;
         let mut path = PathBuf::new();
         path.extend(href_attr.split('/'));
         let asset = AssetFile::new(&html_dir, path).await?;
-        let icon_type = attrs.get("type").map(|x| x.to_owned());
-        let sizes = attrs.get("sizes").map(|x| x.to_owned());
-        Ok(Self { id, cfg, asset, icon_type, sizes })
+        Ok(Self { id, cfg, asset })
     }
 
     /// Spawn the pipeline for this asset type.
@@ -51,47 +45,33 @@ impl Icon {
     #[tracing::instrument(level = "trace", skip(self))]
     async fn run(self) -> Result<TrunkLinkPipelineOutput> {
         let rel_path = crate::common::strip_prefix(&self.asset.path);
-        tracing::info!(path = ?rel_path, "copying & hashing icon");
+        tracing::info!(path = ?rel_path, "copying & hashing manifest");
         let hashed_file_output = self.asset.copy_with_hash(&self.cfg.staging_dist).await?;
-        tracing::info!(path = ?rel_path, "finished copying & hashing icon");
-        Ok(TrunkLinkPipelineOutput::Icon(IconOutput {
+        tracing::info!(path = ?rel_path, "finished copying & hashing manifest");
+        Ok(TrunkLinkPipelineOutput::Manifest(ManifestOutput {
             cfg: self.cfg.clone(),
             id: self.id,
             file: hashed_file_output,
-            icon_type: self.icon_type,
-            sizes: self.sizes,
         }))
     }
 }
 
 /// The output of an Icon build pipeline.
-pub struct IconOutput {
+pub struct ManifestOutput {
     /// The runtime build config.
     pub cfg: Arc<RtcBuild>,
     /// The ID of this pipeline.
     pub id: usize,
     /// Data on the finalized output file.
     pub file: HashedFileOutput,
-    /// Optional type attribute (like type="image/x-icon")
-    pub icon_type: Option<String>,
-    /// Optional sizes attribute
-    pub sizes: Option<String>,
 }
 
-impl IconOutput {
+impl ManifestOutput {
     pub async fn finalize(self, dom: &mut Document) -> Result<()> {
-        let mut opts: Vec<String> = vec![];
-        if let Some(icon_type) = self.icon_type {
-            opts.push(format!("type=\"{}\"", icon_type));
-        }
-        if let Some(sizes) = self.sizes {
-            opts.push(format!("sizes=\"{}\"", sizes));
-        }
         dom.select(&super::trunk_id_selector(self.id)).replace_with_html(format!(
-            r#"<link rel="icon" href="{base}{file}" {optional}/>"#,
+            r#"<link rel="manifest" href="{base}{file}"/>"#,
             base = &self.cfg.public_url,
-            file = self.file.file_name,
-            optional = opts.join(" "),
+            file = self.file.file_name
         ));
         Ok(())
     }